[Refactor]: Split initialize_weights in MAE to submodules
