+++ mmselfsup/models/algorithms/mae.py
@@ -1,9 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-import torch
-from torch import nn
-
 from ..builder import ALGORITHMS, build_backbone, build_head, build_neck
-from ..utils import build_2d_sincos_position_embedding
 from .base import BaseModel
 
 
@@ -30,37 +26,10 @@ class MAE(BaseModel):
         self.neck.num_patches = self.backbone.patch_embed.num_patches
         assert head is not None
         self.head = build_head(head)
-        self.initialize_weights()
 
-    def initialize_weights(self):
-        pos_embed = build_2d_sincos_position_embedding(
-            int(self.backbone.patch_embed.num_patches**.5),
-            self.backbone.pos_embed.shape[-1],
-            cls_token=True)
-        self.backbone.pos_embed.data.copy_(pos_embed.float())
-
-        decoder_pos_embed = build_2d_sincos_position_embedding(
-            int(self.backbone.patch_embed.num_patches**.5),
-            self.neck.decoder_pos_embed.shape[-1],
-            cls_token=True)
-        self.neck.decoder_pos_embed.data.copy_(decoder_pos_embed.float())
-
-        w = self.backbone.patch_embed.projection.weight.data
-        torch.nn.init.xavier_uniform_(w.view([w.shape[0], -1]))
-
-        torch.nn.init.normal_(self.backbone.cls_token, std=.02)
-        torch.nn.init.normal_(self.neck.mask_token, std=.02)
-
-        self.apply(self._init_weights)
-
-    def _init_weights(self, m):
-        if isinstance(m, nn.Linear):
-            torch.nn.init.xavier_uniform_(m.weight)
-            if isinstance(m, nn.Linear) and m.bias is not None:
-                nn.init.constant_(m.bias, 0)
-        elif isinstance(m, nn.LayerNorm):
-            nn.init.constant_(m.bias, 0)
-            nn.init.constant_(m.weight, 1.0)
+    
+    def init_weights(self):
+        super(MAE, self).init_weights()
 
     def extract_feat(self, img):
         """Function to extract features from backbone.
